requirements.txt
moviepy==1.0.3
numpy==1.26.4
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.4.0
